Extract rename modal to rename_modal.rs (refs #68)
Moves RenameTarget, RenameModal, and render_rename_modal() out of
main.rs into a dedicated rename_modal.rs module. Adds Copy/Clone
derives to RenameTarget to simplify the match ownership.

Mirrors wezterm-gui's per-concern module layout under termwindow/.

No behavior change — pure mechanical refactor.
